Drivers/OpTeeRpmb: Add an OP-TEE backed RPMB driver
A following patch is adding support for building StMM in order to run it
from OP-TEE.
OP-TEE in combination with a NS-world supplicant can use the RPMB
partition of an eMMC to store EFI variables. The supplicant
functionality is currently available in U-Boot. Assuming similar
functionality is added in EDK2, this allows any device with an RPMB to
store EFI variables securely.

So let's add a driver that enables access of the  RPMB partition through
OP-TEE.

Signed-off-by: Ilias Apalodimas <ilias.apalodimas@linaro.org>