StMMRpmb: Add support for building StandaloneMm image for OP-TEE
With some recent changes in OP-TEE [1] we can compile StMM and launch it
from an OP-TEE secure partition which is mimicking SPM.

There's a number of advantages in this approach. In Arms world SPM,
currently used for dispatching StMM, and SPD used for OP-TEE, are
mutually exclusive. Since there's no application in OP-TEE for managing
EFI variables, this means that one can have a secure OS or secure
variable storage.

By re-using StMM we have EDK2s approved application controlling
variable storage and the ability to run a secure world OS. This also
allows various firmware implementations to adopt EDK2 way of storing
variables (including the FTW implementation), as long as OP-TEE is
available on that given platform.
Another advantage is that OP-TEE has the ability to access an eMMC RPMB
partition to store those variables. This requires a normal world
supplicant (which is implemented in U-Boot currently) which picks up the
buffer from OP-TEE and wires it to the RPMB driver(s).

There's is one drawback in trying to store data in an RPMB partition
using OP-TEE. The current SPM calls need to run to completion. This
contradicts the current OP-TEE RPC call requirements, used to access
the storage. Thats leads to two different SMC calls for entering secure
world to access StMM. The relevant code for EDK2 can be found here [2]

So let's add support for a platform that compiles StMM and an OP-TEE
supported RPMB driver.

[1] https://github.com/apalos/optee_os/tree/stmm_upstream_03_clean
(not yet upstreamed)
[2] https://github.com/sahilnxp/edk2/pull/1/commits/69aff9e36b770a73f9b2168c8bd413bd696785e8
(will be upstreamed)

Signed-off-by: Ilias Apalodimas <ilias.apalodimas@linaro.org>